better cache

+++ qcpu/src/main.rs
@@ -348,9 +348,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                         stat: sim.stat,
                         const_: Constants {
                             clock_mhz: clock as u64,
-                            cache_hit_penalty: CACHE_HIT_PENALTY,
-                            cache_miss_penalty: CACHE_MISS_PENALTY,
-                            inw_delay: INW_DELAY,
+                            cache_hit_penalty: CACHE_HIT_PENALTY as u64,
+                            cache_miss_penalty: CACHE_MISS_PENALTY as u64,
+                            inw_delay: INW_DELAY as u64,
                         },
                     };
